requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
attrs==25.1.0
audioop-lts==0.2.1
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
discord.py==2.5.2
dnspython==2.7.0
frozenlist==1.5.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
gspread==6.0.2
gspread_asyncio==2.0.0
idna==3.10
multidict==6.1.0
numpy==2.2.2
oauthlib==3.3.1
opencv-python==4.11.0.86
pillow==11.2.1
propcache==0.2.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pygame==2.6.1
pymongo==4.11
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
StrEnum==0.4.15
urllib3==2.5.0
yarl==1.18.3
